apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
PyYAML==6.0.1
python-dateutil==2.8.2
